xchargo: Remove use of packSocketType
Newer versions of the Network.Socket module stop exporting
packSocketType, so it needs to be replaced.

Tools.Argo is only used for Stream sockets, so support for them can be
hardcoded.  Since this is an .hsc file, we can grab SOCK_STREAM from the
C headers and pass it to argo_socket.  Any other type will throw an
error.  Datagram could be supported, but it would be unused and is
skipped for now.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>